clean up gamecontext
